refactor handlers
