+++ localstack/aws/app.py
@@ -1,21 +1,12 @@
 import logging
-import threading
-from typing import Any
 
 from localstack.aws import handlers
-from localstack.http import Response
-from localstack.services.plugins import (
-    SERVICE_PLUGINS,
-    Service,
-    ServiceManager,
-    ServicePluginManager,
-)
+from localstack.aws.handlers.service_plugin import ServiceLoader
+from localstack.services.plugins import SERVICE_PLUGINS, ServiceManager, ServicePluginManager
 
-from .api import RequestContext
-from .chain import HandlerChain
 from .gateway import Gateway
-from .handlers import EmptyResponseHandler, RouterHandler
-from .proxy import AwsApiListener, DefaultListenerHandler, LegacyPluginHandler
+from .handlers.fallback import EmptyResponseHandler
+from .handlers.service import ServiceRequestRouter
 
 LOG = logging.getLogger(__name__)
 
@@ -23,19 +14,12 @@ LOG = logging.getLogger(__name__)
 class LocalstackAwsGateway(Gateway):
     def __init__(self, service_manager: ServiceManager = None) -> None:
         super().__init__()
+
         # basic server components
         self.service_manager = service_manager or ServicePluginManager()
-        self.mutex = threading.RLock()
-
-        # the request router used within the handler chain
-        self.service_request_router = handlers.ServiceRequestRouter()
-
-        # legacy compatibility with DEFAULT_LISTENERS
-        serve_default_listeners = DefaultListenerHandler()
-
-        from localstack.services.edge import ROUTER
-
-        serve_custom_routes = RouterHandler(ROUTER)
+        self.service_request_router = ServiceRequestRouter()
+        # lazy-loads services into the router
+        load_service = ServiceLoader(self.service_manager, self.service_request_router)
 
         # the main request handler chain
         self.request_handlers.extend(
@@ -44,16 +28,16 @@ class LocalstackAwsGateway(Gateway):
                 handlers.parse_service_name,  # enforce_cors and content_decoder depend on the service name
                 handlers.enforce_cors,
                 handlers.content_decoder,
-                handlers.serve_localstack_resources,  # try to serve internal resources first
-                serve_default_listeners,
-                serve_custom_routes,
+                handlers.serve_localstack_resources,  # try to serve internal resources in /_localstack first
+                handlers.serve_default_listeners,  # legacy proxy default listeners
+                handlers.serve_edge_router_rules,
                 # start aws handler chain
                 handlers.inject_auth_header_if_missing,
                 handlers.add_region_from_header,
                 handlers.add_default_account_id,
                 handlers.parse_service_request,
-                self.require_service,
-                self.service_request_router,
+                load_service,  # once we have the service request we can make sure we load the service
+                self.service_request_router,  # once we know the service is loaded we can route the request
                 # if the chain is still running, set an empty response
                 EmptyResponseHandler(404, b'{"message": "Not Found"}'),
             ]
@@ -72,69 +56,10 @@ class LocalstackAwsGateway(Gateway):
         self.response_handlers.extend(
             [
                 handlers.add_cors_response_headers,
-                self.log_response,
+                handlers.log_response,
                 handlers.pop_quart_context,
             ]
         )
-
-    def log_response(self, _: HandlerChain, context: RequestContext, response: Response):
-        if context.operation:
-            # TODO: log analytics event here
-            LOG.info(
-                "%s %s.%s => %d",
-                context.request.method,
-                context.service.service_name,
-                context.operation.name,
-                response.status_code,
-            )
-        else:
-            LOG.info(
-                "%s %s => %d",
-                context.request.method,
-                context.request.path,
-                response.status_code,
-            )
-
-    def require_service(self, _: HandlerChain, context: RequestContext, response: Response):
-        request_router = self.service_request_router
-
-        if not context.service:
-            return
-
-        # verify that we have a route for this request
-        service_operation = context.service_operation
-        if service_operation in request_router.handlers:
-            return
-
-        # FIXME: this blocks all requests to other services, so a mutex list per-service would be useful
-        with self.mutex:
-            # try again to avoid race conditions
-            if service_operation in request_router.handlers:
-                return
-
-            service_name = context.service.service_name
-            if not self.service_manager.exists(service_name):
-                raise NotImplementedError
-
-            service_plugin: Service = self.service_manager.require(service_name)
-
-            if isinstance(service_plugin, Service):
-                if type(service_plugin.listener) == AwsApiListener:
-                    request_router.add_skeleton(service_plugin.listener.skeleton)
-                else:
-                    request_router.add_handler(service_operation, LegacyPluginHandler())
-            else:
-                LOG.warning(
-                    "found plugin for %s, but cannot attach service plugin of type %s",
-                    service_name,
-                    type(service_plugin),
-                )
-
-    def add_provider(self, provider: Any, service_name: str = None):
-        if service_name is None:
-            service_name = provider.service
-
-        self.service_request_router.add_provider(provider=provider, service=service_name)
 
 
 def main():
